Add functions for barres with one count duration
Since the `hbarre` and `fbarre` functions rely on text spans, they don't
take single count values.  Add functions that return markup that
allow this functionality.  The only issue is that the two types are
called differently.  Where the spanning barres are called this way:
```
\hbarre <position> { <music> }
```

The single duration barres must be called like markup:
```
% Fmaj7
<f a c e>^\onefbarre "V"
```

Might try to change this in the future.

Also, rename file from `barre.ly` to `barres.ly`